ASF header added to code and ReadMe.md file updated.
