+++ examples/karaf-redis-example/karaf-redis-example-provider/src/main/java/org/apache/karaf/examples/redis/provider/Activator.java
@@ -1,3 +1,19 @@
+/*
+ *  Licensed to the Apache Software Foundation (ASF) under one or more
+ *  contributor license agreements.  See the NOTICE file distributed with
+ *  this work for additional information regarding copyright ownership.
+ *  The ASF licenses this file to You under the Apache License, Version 2.0
+ *  (the "License"); you may not use this file except in compliance with
+ *  the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 package org.apache.karaf.examples.redis.provider;
 
 import org.apache.karaf.examples.redis.api.UserService;
@@ -15,10 +31,10 @@ public class Activator extends BaseActivator {
 
     @Override
     protected void doStart() throws Exception {
-        UserServiceImp userServiceImp = new UserServiceImp();
-        register(UserService.class, userServiceImp);
-        UserServiceRedisImp userServiceRedisImp = new UserServiceRedisImp();
-        register(UserServiceRedis.class, userServiceRedisImp);
+        UserServiceImpl userServiceImpl = new UserServiceImpl();
+        register(UserService.class, userServiceImpl);
+        UserServiceRedisImpl userServiceRedisImpl = new UserServiceRedisImpl();
+        register(UserServiceRedis.class, userServiceRedisImpl);
     }
 }
