+++ examples/pom.xml
@@ -56,6 +56,7 @@
         <module>karaf-mbean-example</module>
         <module>karaf-profile-example</module>
         <module>karaf-url-namespace-handler-example</module>
+        <module>karaf-redis-example</module>
         <module>karaf-rest-example</module>
         <module>karaf-servlet-example</module>
         <module>karaf-scheduler-example</module>
@@ -63,7 +64,6 @@
         <module>karaf-soap-example</module>
         <module>karaf-websocket-example</module>
         <module>karaf-war-example</module>
-        <module>karaf-redis-example</module>
     </modules>
 
 </project>
